requirements.txt
Flask>=2.2
pandas>=1.5
openpyxl>=3.0.0
xlrd>=2.0.1
reportlab>=3.6.12
arabic-reshaper>=0.3.0
python-bidi>=0.4.2
